# 🧨 Refactor Bazaar: Modular Audio Architecture