--- a/.prettierignore
+++ b/.prettierignore
@@ -1,0 +1,10 @@
+node_modules/
+.next/
+dist/
+build/
+coverage/
+*.min.js
+*.min.css
+*.bundle.js
+*.wasm
+.git/