--- a/components/AudioDemo.module.css
+++ b/components/AudioDemo.module.css
@@ -1,260 +1,48 @@
-/* AudioDemo Component - Modern 2025 CSS Grid Design */
+/* AudioDemo Component - Modern 2025 Clean Design */
 
-/* Main Container - CSS Grid Layout */
-.container {
-  display: grid;
-  grid-template-rows: auto auto 1fr auto;
-  gap: 1.5rem; /* 24px */
-  padding: 2rem; /* 32px */
-  background: rgba(255, 255, 255, 0.08);
-  border-radius: 1rem; /* 16px */
-  border: 1px solid rgba(255, 255, 255, 0.1);
-  backdrop-filter: blur(10px);
-  box-shadow: 0 10px 15px -3px rgba(0, 0, 0, 0.2);
-  min-height: 500px;
+/* Advanced Container - iPhone-like narrow design */
+.advancedContainer {
+  display: flex;
+  flex-direction: column;
+  gap: 1rem;
+  padding: 1rem;
+  background: #0a0a0a;
+  border-radius: 12px;
+  border: 1px solid #1a1a1a;
+  min-height: auto;
+  max-width: 480px;
+  margin: 0 auto;
+  width: 100%;
 }
 
-/* Header Section */
+/* Header Section - Simplified */
 .header {
-  display: grid;
-  grid-template-columns: 1fr auto;
+  display: flex;
+  justify-content: space-between;
   align-items: center;
-  gap: 1rem; /* 16px */
-  padding-bottom: 1.5rem; /* 24px */
-  border-bottom: 1px solid rgba(255, 255, 255, 0.1);
+  padding-bottom: 1rem;
+  border-bottom: 1px solid #1a1a1a;
 }
 
 .header h3 {
   margin: 0;
-  font-size: 1.5rem;
-  font-weight: 600;
-  color: rgba(255, 255, 255, 0.95);
-  display: flex;
-  align-items: center;
-  gap: 1rem; /* 16px */
-}
-
-/* Status Badge */
-.ready,
-.notReady {
-  display: inline-flex;
-  align-items: center;
-  padding: 0.5rem 1rem; /* 8px 16px */
-  border-radius: 0.75rem; /* 12px */
-  font-size: 0.75rem;
-  font-weight: 600;
-  text-transform: uppercase;
-  letter-spacing: 0.05em;
-}
-
-.ready {
-  background: rgba(16, 185, 129, 0.1);
-  color: #10b981;
-  border: 1px solid rgba(16, 185, 129, 0.2);
-}
-
-.notReady {
-  background: rgba(245, 158, 11, 0.1);
-  color: #f59e0b;
-  border: 1px solid rgba(245, 158, 11, 0.2);
-}
-
-/* Error State */
-.error {
-  display: grid;
-  grid-template-columns: auto 1fr;
-  align-items: center;
-  gap: 1rem; /* 16px */
-  padding: 1rem; /* 16px */
-  background: rgba(239, 68, 68, 0.1);
-  border: 1px solid rgba(239, 68, 68, 0.2);
-  border-radius: 0.5rem; /* 8px */
-  color: #ef4444;
-  font-size: 0.875rem;
-}
-
-.error::before {
-  content: "⚠️";
   font-size: 1.25rem;
-}
-
-/* Process Button - Modern Design */
-.processButton {
-  justify-self: center;
-  display: inline-flex;
-  align-items: center;
-  gap: 1rem; /* 16px */
-  padding: 1rem 2rem; /* 16px 32px */
-  background: linear-gradient(135deg, #6366f1 0%, #4f46e5 100%);
-  color: white;
-  border: none;
-  border-radius: 0.75rem; /* 12px */
-  font-size: 1rem;
   font-weight: 600;
-  cursor: pointer;
-  transition: all 0.2s ease;
-  box-shadow: 0 4px 12px rgba(99, 102, 241, 0.3);
+  color: #ffffff;
 }
 
-.processButton:hover:not(:disabled) {
-  transform: translateY(-2px);
-  box-shadow: 0 8px 20px rgba(99, 102, 241, 0.4);
-}
 
-.processButton:active:not(:disabled) {
-  transform: translateY(0);
-}
-
-.processButton:disabled {
-  opacity: 0.6;
-  cursor: not-allowed;
-  box-shadow: none;
-}
-
-/* Audio Players Grid */
-.audioPlayers {
-  display: grid;
-  grid-template-columns: repeat(auto-fit, minmax(280px, 1fr));
-  gap: 1.5rem; /* 24px */
-  padding: 1rem; /* 16px */
-}
-
-/* Audio Card - Modern Card Design */
-.audioCard {
-  display: grid;
-  grid-template-rows: auto 1fr;
-  gap: 1rem; /* 16px */
-  padding: 1.5rem; /* 24px */
-  background: rgba(255, 255, 255, 0.03);
-  border: 1px solid rgba(255, 255, 255, 0.1);
-  border-radius: 0.75rem; /* 12px */
-  transition: all 0.2s ease;
-}
-
-.audioCard:hover {
-  background: rgba(255, 255, 255, 0.12);
-  border-color: rgba(255, 255, 255, 0.2);
-  transform: translateY(-1px);
-  box-shadow: 0 4px 6px -1px rgba(0, 0, 0, 0.1);
-}
-
-.audioCardTitle {
-  margin: 0;
-  font-size: 1.125rem;
-  font-weight: 600;
-  color: rgba(255, 255, 255, 0.95);
-  display: flex;
-  align-items: center;
-  gap: 0.5rem; /* 8px */
-}
-
-.audioCardTitle::before {
-  content: "🎵";
-  font-size: 1rem;
-  opacity: 0.8;
-}
-
-/* Audio Player Styling */
-.audioPlayer {
-  width: 100%;
-  height: 40px;
-  border-radius: 0.375rem; /* 6px */
-  background: rgba(0, 0, 0, 0.2);
-  filter: invert(0.9);
-}
-
-/* Audio Placeholder */
-.audioPlaceholder {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: 100%;
-  height: 40px;
-  background: rgba(255, 255, 255, 0.05);
-  border: 1px dashed rgba(255, 255, 255, 0.2);
-  border-radius: 0.375rem; /* 6px */
-}
-
-.placeholderText {
-  font-size: 0.875rem;
-  color: rgba(255, 255, 255, 0.5);
-  text-align: center;
-}
-
-/* Settings Footer */
-.settings {
-  display: grid;
-  grid-template-columns: repeat(auto-fit, minmax(150px, 1fr));
-  gap: 1rem; /* 16px */
-  padding: 1rem; /* 16px */
-  background: rgba(255, 255, 255, 0.02);
-  border-radius: 0.5rem; /* 8px */
-  border: 1px solid rgba(255, 255, 255, 0.1);
-  font-size: 0.875rem;
-  color: rgba(255, 255, 255, 0.7);
-}
-
-.settings strong {
-  color: rgba(255, 255, 255, 0.95);
-  font-weight: 600;
-}
-
-/* Responsive Design */
-@media (max-width: 640px) {
-  .container {
-    padding: 1rem; /* 16px */
-    gap: 1rem; /* 16px */
-  }
-  
-  .audioPlayers {
-    grid-template-columns: 1fr;
-  }
-  
-  .processButton {
-    width: 100%;
-    justify-content: center;
-  }
-}
-
-/* Animation Keyframes */
-@keyframes fadeIn {
-  from {
-    opacity: 0;
-    transform: translateY(10px);
-  }
-  to {
-    opacity: 1;
-    transform: translateY(0);
-  }
-}
-
-.container {
-  animation: fadeIn 0.3s ease-out;
-}
-
-/* Advanced Container - Multi-panel Layout */
-.advancedContainer {
-  display: grid;
-  grid-template-rows: auto 1fr;
-  gap: 1.5rem;
-  padding: 2rem;
-  background: rgba(255, 255, 255, 0.08);
-  border-radius: 1rem;
-  border: 1px solid rgba(255, 255, 255, 0.1);
-  backdrop-filter: blur(10px);
-  box-shadow: 0 10px 15px -3px rgba(0, 0, 0, 0.2);
-  min-height: 600px;
-  animation: fadeIn 0.3s ease-out;
-}
-
-/* Main Grid - 3 column layout */
+/* Main Grid - Mobile-first single column */
 .mainGrid {
   display: grid;
-  grid-template-columns: 300px 1fr 400px;
-  gap: 1.5rem;
+  grid-template-columns: 1fr;
+  grid-template-rows: auto 1fr auto;
+  gap: 1rem;
   height: 100%;
-  min-height: 500px;
+  min-height: auto;
+  max-width: 100%;
 }
+
 
 /* Header with Actions */
 .headerActions {
@@ -289,11 +77,14 @@
   cursor: not-allowed;
 }
 
-/* Control Panel */
+/* Control Panel - Compact for mobile */
 .controlPanel {
-  display: grid;
-  grid-template-rows: auto auto 1fr auto;
-  gap: 1.5rem;
+  display: flex;
+  flex-direction: column;
+  gap: 0.75rem;
+  padding: 1rem;
+  background: rgba(255, 255, 255, 0.05);
+  border-radius: 0.75rem;
 }
 
 /* Metrics Card */
@@ -311,11 +102,11 @@
   color: rgba(255, 255, 255, 0.95);
 }
 
-/* Metrics Grid */
+/* Metrics Grid - 2x2 for mobile */
 .metricsGrid {
   display: grid;
   grid-template-columns: 1fr 1fr;
-  gap: 0.75rem;
+  gap: 0.5rem;
 }
 
 /* Metric Item */
@@ -381,13 +172,162 @@
   color: rgba(255, 255, 255, 0.95);
 }
 
-/* Audio Panel */
+/* Audio Panel - MAIN FOCUS */
 .audioPanel {
-  display: grid;
-  align-content: start;
+  display: flex;
+  flex-direction: column;
   gap: 1rem;
   overflow-y: auto;
   max-height: 100%;
+  background: rgba(0, 0, 0, 0.5);
+  border: 2px solid rgba(102, 126, 234, 0.3);
+  border-radius: 1rem;
+  padding: 1.5rem;
+  min-height: 400px;
+}
+
+/* Audio Players Container */
+.audioPlayers {
+  display: grid;
+  grid-template-columns: 1fr;
+  gap: 1rem;
+}
+
+/* Audio Card */
+.audioCard {
+  background: rgba(255, 255, 255, 0.05);
+  border: 1px solid rgba(255, 255, 255, 0.1);
+  border-radius: 0.75rem;
+  padding: 1.25rem;
+  transition: all 0.2s ease;
+}
+
+.audioCard:hover {
+  background: rgba(255, 255, 255, 0.08);
+  border-color: rgba(255, 255, 255, 0.2);
+}
+
+.audioCardTitle {
+  margin: 0 0 0.75rem 0;
+  font-size: 1rem;
+  font-weight: 600;
+  color: rgba(255, 255, 255, 0.95);
+}
+
+/* Audio Player */
+.audioPlayer {
+  width: 100%;
+  height: 40px;
+  border-radius: 0.5rem;
+  background: rgba(0, 0, 0, 0.3);
+}
+
+/* Audio Placeholder */
+.audioPlaceholder {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  height: 40px;
+  background: rgba(255, 255, 255, 0.03);
+  border: 1px dashed rgba(255, 255, 255, 0.2);
+  border-radius: 0.5rem;
+}
+
+.placeholderText {
+  font-size: 0.875rem;
+  color: rgba(255, 255, 255, 0.5);
+  text-align: center;
+}
+
+/* Audio Panel Title */
+.audioPanelTitle {
+  margin: 0 0 1rem 0;
+  font-size: 1.25rem;
+  font-weight: 600;
+  color: rgba(255, 255, 255, 0.95);
+  text-align: center;
+}
+
+/* Bottom Panel */
+.bottomPanel {
+  display: flex;
+  flex-direction: column;
+  gap: 1rem;
+}
+
+/* Settings */
+.settings {
+  display: flex;
+  justify-content: center;
+  gap: 1rem;
+  padding: 0.75rem;
+  background: rgba(255, 255, 255, 0.03);
+  border-radius: 0.5rem;
+  font-size: 0.875rem;
+  color: rgba(255, 255, 255, 0.8);
+}
+
+/* Process Button */
+.processButton {
+  padding: 1rem 2rem;
+  background: linear-gradient(135deg, #667eea 0%, #764ba2 100%);
+  color: white;
+  border: none;
+  border-radius: 0.75rem;
+  font-size: 1rem;
+  font-weight: 600;
+  cursor: pointer;
+  transition: all 0.2s ease;
+  width: 100%;
+}
+
+.processButton:hover:not(:disabled) {
+  transform: translateY(-2px);
+  box-shadow: 0 4px 12px rgba(102, 126, 234, 0.4);
+}
+
+.processButton:disabled {
+  opacity: 0.5;
+  cursor: not-allowed;
+  transform: none;
+}
+
+/* Details Elements */
+.chunkDetails,
+.logsDetails {
+  background: rgba(255, 255, 255, 0.05);
+  border: 1px solid rgba(255, 255, 255, 0.1);
+  border-radius: 0.75rem;
+  padding: 0;
+  overflow: hidden;
+}
+
+.chunkDetails summary,
+.logsDetails summary {
+  padding: 1rem;
+  cursor: pointer;
+  font-weight: 600;
+  color: rgba(255, 255, 255, 0.9);
+  background: rgba(255, 255, 255, 0.03);
+  border-radius: 0.75rem;
+  user-select: none;
+  list-style: none;
+}
+
+.chunkDetails summary::-webkit-details-marker,
+.logsDetails summary::-webkit-details-marker {
+  display: none;
+}
+
+.chunkDetails[open] summary,
+.logsDetails[open] summary {
+  border-radius: 0.75rem 0.75rem 0 0;
+  border-bottom: 1px solid rgba(255, 255, 255, 0.1);
+}
+
+.chunkDetails[open] .chunkList,
+.logsDetails[open] .logsContainer {
+  padding: 1rem;
 }
 
 /* Audio Metrics */
@@ -513,26 +453,91 @@
   background: rgba(255, 255, 255, 0.3);
 }
 
-/* Responsive adjustments */
-@media (max-width: 1400px) {
-  .mainGrid {
-    grid-template-columns: 250px 1fr 350px;
-  }
-}
-
-@media (max-width: 1200px) {
-  .mainGrid {
-    grid-template-columns: 1fr;
-    grid-template-rows: auto auto auto;
+/* Enhanced mobile styles */
+@media (max-width: 480px) {
+  .advancedContainer {
+    padding: 0.75rem;
+    border-radius: 0;
+    border-left: none;
+    border-right: none;
   }
   
-  .controlPanel,
-  .audioPanel,
-  .logsPanel {
-    max-height: none;
+  .audioPanel {
+    padding: 1rem;
+    border-radius: 0.75rem;
+  }
+  
+  .metricsGrid {
+    font-size: 0.875rem;
   }
   
   .logsContainer {
-    max-height: 300px;
+    max-height: 200px;
+    font-size: 0.7rem;
   }
+}
+
+/* Tablet and larger screens */
+@media (min-width: 768px) {
+  .advancedContainer {
+    max-width: 600px;
+    padding: 1.5rem;
+  }
+  
+  .audioPlayers {
+    grid-template-columns: repeat(auto-fit, minmax(250px, 1fr));
+  }
+}
+
+/* Status indicators */
+.ready,
+.notReady {
+  padding: 0.25rem 0.75rem;
+  border-radius: 0.375rem;
+  font-size: 0.75rem;
+  font-weight: 500;
+  text-transform: uppercase;
+  letter-spacing: 0.05em;
+}
+
+.ready {
+  background: rgba(34, 197, 94, 0.2);
+  color: #22c55e;
+  border: 1px solid rgba(34, 197, 94, 0.3);
+}
+
+.notReady {
+  background: rgba(239, 68, 68, 0.2);
+  color: #ef4444;
+  border: 1px solid rgba(239, 68, 68, 0.3);
+}
+
+/* Error state */
+.error {
+  background: rgba(239, 68, 68, 0.1);
+  border: 1px solid rgba(239, 68, 68, 0.3);
+  color: #ef4444;
+  padding: 1rem;
+  border-radius: 0.5rem;
+  margin-bottom: 1rem;
+  font-size: 0.875rem;
+  line-height: 1.5;
+}
+
+/* Reset button */
+.resetButton {
+  margin-top: 0.5rem;
+  padding: 0.5rem 1rem;
+  background: #f59e0b;
+  color: white;
+  border: none;
+  border-radius: 0.375rem;
+  cursor: pointer;
+  font-size: 0.875rem;
+  font-weight: 500;
+  transition: all 0.2s ease;
+}
+
+.resetButton:hover {
+  background: #d97706;
 }