… .claude/settings.local.json
… .env
… .eslintrc.json
… .prettierrc.json
… BUILD_REPORT.md
… ESLINT_FIXES_SUMMARY.md
… README.md
… components/MurmurabaSuiteStatus.tsx
… components/ui/ControlPanel.tsx
… context/MediaStreamContext.tsx
… docs/ANTI_PATTERNS.md
… docs/CRITICAL-ARCHITECTURE-VIOLATIONS.md
… docs/MODULAR_ARCHITECTURE.md
… hooks/useAudioProcessor.ts
… hooks/useAudioProcessorSuite.ts
… next.config.js
… package.json
… packages/murmuraba/.claude/settings.local.json
… packages/murmuraba/.eslintrc.json
… packages/murmuraba/CHANGELOG.md
… packages/murmuraba/DI_PATTERN_GUIDE.md
… packages/murmuraba/README-ARCHITECTURE.md
… packages/murmuraba/README.md
… packages/murmuraba/docs/CONFIG_VALIDATION.md
… packages/murmuraba/docs/DOM_INTEGRATION_TESTS.md
… packages/murmuraba/docs/MODERN_ARCHITECTURE.md
… packages/murmuraba/examples/config-validation-example.ts
… packages/murmuraba/global.d.ts
… packages/murmuraba/html/assets/index-D_ryMEPs.js
… packages/murmuraba/html/assets/index-X8b7Z_4p.css